Moved sum() to arithmetics, removed reductions.py entirely
